minimal change test=huawei_ascend_npu test=nvidia_tensorrt test=verisilicon_timvx
