+++ lite/api/paddle_use_passes.h
@@ -68,17 +68,16 @@ USE_MIR_PASS(lite_inplace_fuse_pass);
 USE_MIR_PASS(multi_stream_analysis_pass);
 USE_MIR_PASS(elementwise_mul_constant_eliminate_pass);
 USE_MIR_PASS(npu_subgraph_pass);
+USE_MIR_PASS(nnadapter_subgraph_pass);
 USE_MIR_PASS(mlu_subgraph_pass);
 USE_MIR_PASS(mlu_postprocess_pass);
 USE_MIR_PASS(weight_quantization_preprocess_pass);
 USE_MIR_PASS(post_quant_dynamic_pass);
 USE_MIR_PASS(fp16_attribute_pass);
 USE_MIR_PASS(fpga_concat_fuse_pass);
-USE_MIR_PASS(nnadapter_op_fusion_pass);
-USE_MIR_PASS(nnadapter_quantization_parameters_propagation_pass);
-USE_MIR_PASS(nnadapter_restrict_quantized_op_with_same_input_output_scale_pass);
-USE_MIR_PASS(nnadapter_quantization_parameters_removal_pass);
-USE_MIR_PASS(nnadapter_subgraph_pass);
+USE_MIR_PASS(quantization_parameters_propagation_pass);
+USE_MIR_PASS(quantization_parameters_removal_pass);
+USE_MIR_PASS(restrict_quantized_op_with_same_input_output_scale_pass);
 USE_MIR_PASS(control_flow_op_unused_inputs_and_outputs_eliminate_pass);
 USE_MIR_PASS(control_flow_op_shared_inputs_and_outputs_place_sync_pass);
 USE_MIR_PASS(lite_scale_activation_fuse_pass);
@@ -119,3 +118,4 @@ USE_MIR_PASS(fill_constant_calc_offline_pass);
 USE_MIR_PASS(unsqueeze_calc_offline_pass);
 USE_MIR_PASS(scale_calc_offline_pass);
 USE_MIR_PASS(keepdims_convert_pass);
+USE_MIR_PASS(op_fusion_minimal_set_pass);
