+++ lite/core/optimizer/mir/restrict_quantized_op_with_same_input_output_scale_pass.h
@@ -68,8 +68,7 @@ namespace mir {
 //                           |
 //                           v
 
-class NNAdapterRestrictQuantizedOpWithSameInputOutputScalePass
-    : public mir::StmtPass {
+class RestrictQuantizedOpWithSameInputOutputScalePass : public mir::StmtPass {
  public:
   void Apply(const std::unique_ptr<SSAGraph>& graph) override;
 };
